Version 1.0.19
Fixed implementation of config overrides system, where overrides now "survive" a config reload.
Also, if a config_overrides_file is provided, this is also polled for changes, and will trigger a reload.
Config overrides file is now non-fatal, in that a missing file or JSON errors will not cause a crash.
When calling logDebug(), logError() or logTransaction() before a logger is initialized, these now silently fail (instead of crash).
Bumped pixl-config to ^1.0.7 for bug fix.